+++ packages/analytics/dashboard-renderer/sandbox/sandbox-query-provider.ts
@@ -56,7 +56,7 @@ const configFn = (): Promise<AnalyticsConfigV2> => {
 
 const evaluateFeatureFlagFn = () => true
 
-const exploreBaseUrl = () => 'https://cloud.konghq.tech/us/analytics/explorer'
+const exploreBaseUrl = async () => 'https://cloud.konghq.tech/us/analytics/explorer'
 
 const fetchComponent = async (name: string): Promise<Component> => {
   return Promise.resolve(EntityLink)