+++ packages/entities/entities-plugins/src/components/PluginEntityForm.vue
@@ -6,12 +6,35 @@
     />
 
     <div
-      v-else
+      v-else-if="(formModel.id && editing) || !editing"
       class="entity-form"
     >
       <component
+        :is="freeformName"
+        v-if="freeformName"
+        :form-model="formModel"
+        :form-options="formOptions"
+        :form-schema="formSchema"
+        :is-editing="editing"
+        :model="record"
+        :on-config-change="updateConfig"
+        :on-model-updated="onModelUpdated"
+        :schema="rawSchema"
+      >
+        <template
+          v-if="enableVaultSecretPicker"
+          #[AUTOFILL_SLOT_NAME]="slotProps: AutofillSlotProps"
+        >
+          <VaultSecretPickerProvider
+            v-if="slotProps.schema.referenceable"
+            v-bind="slotProps"
+            @open="setUpVaultSecretPicker"
+          />
+        </template>
+      </component>
+      <component
         :is="sharedFormName"
-        v-if="sharedFormName && (formModel.id && editing || !editing)"
+        v-else-if="sharedFormName"
         :enable-redis-partial="enableRedisPartial"
         :form-model="formModel"
         :form-options="formOptions"
@@ -28,13 +51,13 @@
           <VaultSecretPickerProvider
             v-if="slotProps.schema.referenceable"
             v-bind="slotProps"
-            @open="(value, update) => setUpVaultSecretPicker(value, update)"
+            @open="setUpVaultSecretPicker"
           />
         </template>
       </component>
 
       <VueFormGenerator
-        v-if="!sharedFormName && (formModel.id && editing || !editing)"
+        v-else
         :enable-redis-partial="enableRedisPartial"
         :is-editing="editing"
         :model="formModel"
@@ -66,7 +89,7 @@
           <VaultSecretPickerProvider
             v-if="slotProps.schema.referenceable"
             v-bind="slotProps"
-            @open="(value, update) => setUpVaultSecretPicker(value, update)"
+            @open="setUpVaultSecretPicker"
           />
         </template>
       </VueFormGenerator>
@@ -99,21 +122,32 @@ import {
 } from '@kong-ui-public/forms'
 import '@kong-ui-public/forms/dist/style.css'
 import type { AxiosRequestConfig, AxiosResponse } from 'axios'
-import { computed, defineComponent, onBeforeMount, provide, reactive, ref, watch, type PropType } from 'vue'
+import { computed, defineComponent, onBeforeMount, provide, reactive, ref, shallowRef, watch, type PropType } from 'vue'
 import composables from '../composables'
 import useI18n from '../composables/useI18n'
 import { PLUGIN_METADATA } from '../definitions/metadata'
 import endpoints from '../plugins-endpoints'
-import {
-  type KongManagerPluginFormConfig,
-  type KonnectPluginFormConfig,
-  type PluginEntityInfo,
-} from '../types'
+import type { KongManagerPluginFormConfig, KonnectPluginFormConfig, PluginEntityInfo } from '../types'
 import PluginFieldRuleAlerts from './PluginFieldRuleAlerts.vue'
+import * as freeForm from './free-form'
+import { getFreeFormName } from '../utils/free-form'
+
+// Need to check for duplicates in sharedForms and freeForm
+// throw an error if there are any
+const sharedFormKeys = Object.keys(sharedForms)
+const freeFormKeys = Object.keys(freeForm)
+if (
+  new Set([...sharedFormKeys, ...freeFormKeys]).size !==
+  sharedFormKeys.length + freeFormKeys.length
+) {
+  throw new Error(
+    'Duplicate form component names found in `sharedForms` and `freeForm`',
+  )
+}
 
 // Must explicitly specify these as components since they are rendered dynamically
 export default defineComponent({
-  components: { ...sharedForms },
+  components: { ...sharedForms, ...freeForm },
 })
 </script>
 
@@ -164,6 +198,13 @@ const props = defineProps({
     default: () => ({}),
   },
   /**
+   * Raw schema
+   */
+  rawSchema: {
+    type: Object as PropType<Record<string, any>>,
+    default: () => ({}),
+  },
+  /**
    * Whether or not the form is in edit mode
    */
   editing: {
@@ -195,7 +236,7 @@ const props = defineProps({
 
 const { axiosInstance } = useAxios(props.config?.axiosRequestConfig)
 
-const { parseSchema } = composables.useSchemas({
+const { parseSchema, pruneRecord } = composables.useSchemas({
   entityId: props.entityMap.focusedEntity?.id || undefined,
   credential: props.credential,
   enableRedisPartial: props.enableRedisPartial,
@@ -293,6 +334,7 @@ provide(FORMS_API_KEY, {
 provide(FORMS_CONFIG, props.config)
 
 const sharedFormName = ref('')
+const freeformName = ref('')
 const form = ref<Record<string, any> | null>(null)
 const formSchema = ref<Record<string, any>>({})
 const originalModel = reactive<Record<string, any>>({})
@@ -528,7 +570,15 @@ const getModel = (): Record<string, any> => {
     unsetNullForeignKey(fieldNameDotNotation, outputModel)
   })
 
-  return unFlattenObject(outputModel)
+  const model: Record<string, any> = unFlattenObject(outputModel)
+
+  // Handle the special case of the freeform plugin
+  if (freeformName.value) {
+    const configSchema = props.rawSchema?.fields?.find((field: any) => 'config' in field)?.config
+    model.config = pruneRecord(freeformConfig.value, configSchema)
+  }
+
+  return model
 }
 
 const onPartialToggled = (dismissSchemaField: string | undefined, additionalModel: Record<string, any> = {}) => {
@@ -626,6 +676,19 @@ const updateModel = (data: Record<string, any>, parent?: string) => {
   })
 }
 
+const freeformConfig = shallowRef<Record<string, any>>(props.record.config)
+const updateConfig = (value: Record<string, any>) => {
+  freeformConfig.value = value
+
+  emit('model-updated', {
+    // config change should also update the form model
+    // otherwise the submit button will be disabled
+    model: { ...formModel, config: value },
+    originalModel,
+    data: getModel(),
+  })
+}
+
 const loading = ref(true)
 const initFormModel = (): void => {
   if (props.record && props.schema) {
@@ -665,7 +728,12 @@ const initFormModel = (): void => {
       }
 
       // main plugin configuration
-      updateModel(props.record.config, 'config')
+      if (freeformName.value) {
+        // keep original config from record for freeform plugins
+        updateConfig(props.record.config)
+      } else {
+        updateModel(props.record.config, 'config')
+      }
     }
   }
 
@@ -721,6 +789,8 @@ watch(() => props.schema, (newSchema, oldSchema) => {
     }),
   }
   Object.assign(originalModel, JSON.parse(JSON.stringify(form.model)))
+
+  freeformName.value = getFreeFormName(form.model.name)
   sharedFormName.value = getSharedFormName(form.model.name)
 
   initFormModel()
@@ -760,7 +830,7 @@ onBeforeMount(() => {
   }
 
   :deep(.vue-form-generator) {
-    >fieldset {
+    > fieldset {
       .form-group:last-child {
         margin-bottom: 0;
       }
@@ -774,6 +844,7 @@ onBeforeMount(() => {
 
     fieldset {
       border: none;
+      margin-left: $kui-space-0;
       padding: $kui-space-0;
     }
 
@@ -793,12 +864,12 @@ onBeforeMount(() => {
 
     .form-group hr.divider {
       border-color: $kui-color-border;
-      opacity: .3;
+      opacity: 0.3;
     }
 
     .form-group hr.wide-divider {
       border-color: $kui-color-border;
-      opacity: .6;
+      opacity: 0.6;
     }
 
     .form-group.field-textArea textarea {
@@ -823,7 +894,7 @@ onBeforeMount(() => {
       }
     }
 
-    .field-radios .radio-list label input[type=radio] {
+    .field-radios .radio-list label input[type="radio"] {
       margin-right: 10px;
     }