+++ packages/analytics/portal-analytics-bridge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.77](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.76...@kong-ui-public/portal-analytics-bridge@0.2.77) (2025-03-27)
+
+**Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge
+
+
+
+
+
 ## [0.2.76](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/portal-analytics-bridge@0.2.75...@kong-ui-public/portal-analytics-bridge@0.2.76) (2025-03-20)
 
 **Note:** Version bump only for package @kong-ui-public/portal-analytics-bridge