+++ packages/core/forms/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.12.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.11.17...@kong-ui-public/forms@4.12.0) (2025-04-02)
+
+
+### Features
+
+* **vfg:** introduce custom field component ([#2060](https://github.com/Kong/public-ui-components/issues/2060)) ([1b49948](https://github.com/Kong/public-ui-components/commit/1b49948fde5e5eb0c97b05d8568c911b98656a4e))
+
+
+
+
+
+## [4.11.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.11.16...@kong-ui-public/forms@4.11.17) (2025-03-31)
+
+**Note:** Version bump only for package @kong-ui-public/forms
+
+
+
+
+
 ## [4.11.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.11.15...@kong-ui-public/forms@4.11.16) (2025-03-27)