+++ packages/entities/entities-data-plane-nodes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.87](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.3.86...@kong-ui-public/entities-data-plane-nodes@0.3.87) (2025-03-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes
+
+
+
+
+
 ## [0.3.86](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.3.85...@kong-ui-public/entities-data-plane-nodes@0.3.86) (2025-03-26)
 
 **Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes